+++ db/migrate/20260725120000_add_onboarding_required_to_installations.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddOnboardingRequiredToInstallations < ActiveRecord::Migration[8.0]
+  def change
+    add_column :installations, :onboarding_required, :boolean, default: false, null: false
+  end
+end